+++ src/App.tsx
@@ -4,7 +4,7 @@ import GeographicPrevalence from "@/sections/GeographicPrevalence";
 import Header from "@/sections/Header";
 import Overview from "@/sections/Overview";
 import Recipes from "@/sections/Recipes";
-import SearchMeta from "@/sections/SearchMeta";
+import Search from "@/sections/Search";
 import TaxonomicPrevalence from "@/sections/TaxonomicPrevalence";
 import "@/components/tooltip";
 import "./App.css";
@@ -16,7 +16,7 @@ const App = () => (
     <Header />
     <main>
       <Overview />
-      <SearchMeta />
+      <Search />
       <TaxonomicPrevalence />
       <GeographicPrevalence />
       <Recipes />
