+++ src/sections/Search.tsx
@@ -1,128 +1,47 @@
-import { useEffect, useMemo, useState } from "react";
-import Button from "@/components/Button";
-import Select from "@/components/Select";
-import Textbox from "@/components/Textbox";
-import { Data, SearchList, useData } from "@/data";
-import { thread } from "@/workers";
-import classes from "./Search.module.css";
+import Tabs from "@/components/Tabs";
+import SearchBox from "@/sections/SearchBox";
 
-type Filters = NonNullable<Data["searchList"]>[number]["type"][];
-type FiltersAll = ("All" | Props["filters"][number])[];
+const Search = () => (
+  <section>
+    <h2>Search</h2>
 
-type Props = {
-  filters: Filters;
-};
-
-const Search = ({ filters }: Props) => {
-  const [search, setSearch] = useState("");
-  const [fuzzy, setFuzzy] = useState<SearchList>([]);
-  const [filter, setFilter] = useState<FiltersAll[number]>("All");
-  const [limit, setLimit] = useState(10);
-
-  const fullSearchList = useData((state) => state.searchList);
-
-  /** filter full search list before any other steps */
-  const searchList = useMemo(
-    () =>
-      fullSearchList?.filter((entry) =>
-        entry.type === filter || filter === "All"
-          ? filters.includes(entry.type)
-          : false,
-      ),
-    [filters, filter, fullSearchList],
-  );
-
-  /** get exact matches */
-  const exact = (searchList || []).filter((entry) =>
-    entry.name.toLowerCase().includes(search.toLowerCase()),
-  );
-
-  /** get fuzzy (trigram) matches (in worker to not freeze ui) */
-  useEffect(() => {
-    let latest = true;
-
-    /** reset fuzzy */
-    setFuzzy([]);
-
-    if (searchList && search.trim())
-      /** fuzzy search then set recommendations */
-      thread((worker) => worker.fuzzySearch(searchList, "name", search)).then(
-        (fuzzy) => {
-          /** if not latest run of use effect (superseded), ignore result */
-          if (latest) setFuzzy(fuzzy as SearchList);
+    <Tabs
+      tabs={[
+        {
+          name: "Metadata",
+          description: (
+            <p>
+              Search for a{" "}
+              <span data-tooltip="Collection of multiple samples, by <a href='https://www.ncbi.nlm.nih.gov/bioproject/' target='_blank'>BioProject</a> accession.">
+                project
+              </span>
+              ,{" "}
+              <span data-tooltip="Individual sample, by <a href='https://www.ncbi.nlm.nih.gov/sra' target='_blank'>SRA</a> run accession.">
+                sample
+              </span>
+              ,{" "}
+              <span data-tooltip="Geographic origin of samples, based on <a href='https://www.naturalearthdata.com/' target='_blank'>Natural Earth</a> country data.">
+                country
+              </span>
+              , or{" "}
+              <span data-tooltip="Geographic origin of samples, grouped into regions according to <a href='https://unstats.un.org/sdgs/indicators/regional-groups/' target='_blank'>the UN's Sustainable Development Goals</a>.">
+                region
+              </span>{" "}
+              in the dataset.
+            </p>
+          ),
+          content: (
+            <SearchBox filters={["Project", "Sample", "Country", "Region"]} />
+          ),
         },
-      );
-    else setFuzzy([]);
-
-    return () => {
-      latest = false;
-    };
-  }, [searchList, search]);
-
-  /** full list of matches */
-  const matches = exact.concat(
-    /** de-duplicate items already in exact */
-    fuzzy
-      .filter((fuzzy) => !exact.find((result) => result.name === fuzzy.name))
-      .map((fuzzy) => ({ ...fuzzy, fuzzy: true })),
-  );
-
-  /** reset limit when search changes */
-  useEffect(() => {
-    setLimit(10);
-  }, [search]);
-
-  return (
-    <>
-      <div className={classes.search}>
-        <Textbox value={search} onChange={setSearch} placeholder="Search" />
-        <Select
-          label="Filter by type:"
-          options={["All", ...filters] as FiltersAll}
-          value={filter}
-          onChange={setFilter}
-        />
-      </div>
-
-      <div className="table-wrapper">
-        <table className={classes.table}>
-          <thead>
-            <tr>
-              <th>Name</th>
-              <th>Type</th>
-              <th>Samples</th>
-            </tr>
-          </thead>
-
-          <tbody>
-            <tr>
-              <td colSpan={3}>{/* spacer */}</td>
-            </tr>
-
-            {matches.length ? (
-              matches
-                .slice(0, limit)
-                .map(({ name, type, samples, fuzzy }, index) => (
-                  <tr key={index} style={{ opacity: fuzzy ? 0.75 : 1 }}>
-                    <td>{name}</td>
-                    <td>{type}</td>
-                    <td>{samples.toLocaleString()}</td>
-                  </tr>
-                ))
-            ) : (
-              <tr>
-                <td colSpan={3}>No results</td>
-              </tr>
-            )}
-          </tbody>
-        </table>
-      </div>
-
-      {matches.length > limit && (
-        <Button onClick={() => setLimit(limit + 10)}>Show More</Button>
-      )}
-    </>
-  );
-};
+        {
+          name: "Taxa",
+          description: <p>Search for a phylum or class in the dataset.</p>,
+          content: <SearchBox filters={["Phylum", "Class"]} />,
+        },
+      ]}
+    />
+  </section>
+);
 
 export default Search;
